Create folder for lambda instance pks, refactor fokia
